tools/transcript_asr/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
anyio==4.12.1
attrs==25.4.0
av==16.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
ctranslate2==4.7.1
datasets==4.6.1
dill==0.4.0
faster-whisper==1.2.1
filelock==3.25.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
Jinja2==3.1.6
jiwer==4.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
networkx==3.6.1
numpy==2.4.2
onnxruntime==1.24.2
packaging==26.0
pandas==3.0.1
pillow==12.1.1
propcache==0.4.1
protobuf==7.34.0
pyarrow==23.0.1
pycparser==3.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
RapidFuzz==3.14.3
requests==2.32.5
rich==14.3.3
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
soundfile==0.13.1
sympy==1.14.0
tokenizers==0.22.2
torch==2.10.0
torchaudio==2.10.0
torchcodec==0.10.0
torchvision==0.25.0
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
xxhash==3.6.0
yarl==1.23.0
fastapi
uvicorn
python-multipart
faster-whisper
python-docx
